@@ -3,37 +3,43 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
+    "conventional-changelog": "conventional-changelog",
     "start": "ng serve",
     "build": "ng build",
-    "build:prod": "ng build --prod --optimization",
+    "build:prod": "npm run build -- --prod --aot --optimization",
     "test": "ng test",
+    "test:coverage": "rimraf coverage && npm run test -- --code-coverage",
     "lint": "ng lint",
+    "lint:fix": "ng lint ngx-admin-demo --fix",
+    "lint:styles": "stylelint ./src/**/*.scss",
+    "lint:ci": "npm run lint && npm run lint:styles",
+    "pree2e": "webdriver-manager update --standalone false --gecko false",
     "e2e": "ng e2e",
-    "dev": "lite-server",
-    "compile:server": "webpack --config webpack.server.config.js --progress --colors",
-    "serve:ssr": "node dist/server",
-    "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
-    "build:client-and-server-bundles": "ng build --prod && ng run Amblema-BackOffice:server:production --bundleDependencies all"
+    "docs": "compodoc -p src/tsconfig.app.json -d docs",
+    "docs:serve": "compodoc -p src/tsconfig.app.json -d docs -s",
+    "prepush": "npm run lint:ci",
+    "release:changelog": "npm run conventional-changelog -- -p angular -i CHANGELOG.md -s",
+    "postinstall": "ngcc --properties es2015 es5 browser module main --first-only --create-ivy-entry-points --tsconfig \"./src/tsconfig.app.json\""
   },
   "private": true,
   "dependencies": {
     "@agm/core": "^1.1.0",
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/platform-server": "~8.2.14",
-    "@angular/router": "~8.2.14",
+    "@angular/animations": "^10.2.4",
+    "@angular/cdk": "^10.2.7",
+    "@angular/common": "~10.2.4",
+    "@angular/compiler": "~10.2.4",
+    "@angular/core": "~10.2.4",
+    "@angular/forms": "~10.2.4",
+    "@angular/platform-browser": "~10.2.4",
+    "@angular/platform-browser-dynamic": "~10.2.4",
+    "@angular/platform-server": "~10.2.4",
+    "@angular/router": "~10.2.4",
     "@fortawesome/fontawesome-free": "^5.12.0",
-    "@nebular/auth": "^4.6.0",
+    "@nebular/auth": "^6.0.0",
     "@nebular/date-fns": "^4.6.0",
     "@nebular/eva-icons": "^4.6.0",
     "@nebular/moment": "^4.6.0",
-    "@nebular/theme": "^4.6.0",
+    "@nebular/theme": "^6.0.0",
     "@ng-select/ng-select": "^3.7.3",
     "@nguniversal/express-engine": "^8.2.6",
     "@nguniversal/module-map-ngfactory-loader": "v8.2.6",
@@ -97,11 +104,13 @@
     "karma-jasmine-html-reporter": "^1.4.0",
     "karma-spec-reporter": "0.0.32",
     "lite-server": "^2.5.4",
-    "protractor": "~5.4.0",
+    "patch-package": "^6.2.2",
+    "protractor": "~7.0.0",
+    "stylelint": "^13.8.0",
     "ts-loader": "^5.2.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.5.3",
+    "typescript": "~4.0.5",
     "webpack-cli": "^3.1.0"
   }
 }
