@@ -24,7 +24,8 @@ const routes: Routes = [
 ];
 
 const config: ExtraOptions = {
-  useHash: false,
+    useHash: false,
+    relativeLinkResolution: 'legacy'
 };
 
 @NgModule({
