@@ -1,6 +1,7 @@
 import { AbstractReactive } from '../abstract-reactive';
-import { Input } from '@angular/core';
+import { Input, Directive } from '@angular/core';
 
+@Directive()
 export abstract class AbstractReactiveInput extends AbstractReactive {
     @Input() type: string | null = '';
     @Input() max: number | null = null;
