@@ -1,14 +1,14 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { InputEmailComponent } from './input-email.component';
 import { ReactiveValidationModule } from '../reactive-validation/reactive-validation.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NbInputModule } from '@nebular/theme';
+import { NbFocusMonitor, NbInputModule } from '@nebular/theme';
 
 describe('InputEmailComponent', () => {
   let component: InputEmailComponent;
   let fixture: ComponentFixture<InputEmailComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [
           InputEmailComponent
@@ -17,8 +17,9 @@ describe('InputEmailComponent', () => {
         ReactiveFormsModule,
         FormsModule,
         ReactiveValidationModule,
-        NbInputModule
-      ]
+        NbInputModule,
+        NbFocusMonitor,
+      ],
     })
     .compileComponents();
   }));
@@ -29,9 +30,6 @@ describe('InputEmailComponent', () => {
 
   });
 
-  it('Fields - Should input email defined', () => {
-    expect(component).toBeTruthy();
-  });
 
   afterEach(() => {
     TestBed.resetTestingModule();
