@@ -50,7 +50,7 @@ export class SchoolAdminComponent implements OnInit {
 
       if ( response ) {
         this.timeOut = Date.parse( response.endDate ) < Date.parse( new Date( ).toString() ) ? true : false;
-        console.log( this.timeOut )
+        console.log( this.timeOut );
       }
 
     });
