@@ -30,7 +30,7 @@ export class SponsorComponent implements OnInit {
   ) { }
 
   ngOnInit() {
-  
+
   }
 
 }
