@@ -1,4 +1,4 @@
-import { Inject } from '@angular/core';
+import { Inject, Injectable } from '@angular/core';
 import { DOCUMENT, DatePipe } from '@angular/common';
 import { IMAGE } from '../img-base-64';
 import { OlympicsReport } from 'src/app/_models/report/math-olympics-report.model';
@@ -6,6 +6,7 @@ const pdfMake = require('pdfmake/build/pdfmake.js');
 const pdfFonts = require('pdfmake/build/vfs_fonts.js');
 pdfMake.vfs = pdfFonts.pdfMake.vfs;
 
+@Injectable()
 export class PDFReportMath {
 
 
