@@ -1,17 +1,17 @@
-import { Component, ElementRef } from "@angular/core";
-import { InformationDetailsComponent } from "../information-details/information-details.component";
-import { AuthService } from "src/app/services/user/auth.service";
-import { ALL_ACTIONS } from "src/app/store/_shader/all-actions";
-import { ChartDataSets } from "chart.js";
-import { Label } from "ng2-charts";
-import { InformationRequestService } from "src/app/services/request/information-request.service";
-import { BsModalRef, BsModalService } from "ngx-bootstrap/modal";
-import { Store } from "@ngxs/store";
-import { CustomToastrService } from "src/app/services/helper/custom-toastr.service";
+import { Component, ElementRef } from '@angular/core';
+import { InformationDetailsComponent } from '../information-details/information-details.component';
+import { AuthService } from 'src/app/services/user/auth.service';
+import { ALL_ACTIONS } from 'src/app/store/_shader/all-actions';
+import { ChartDataSets } from 'chart.js';
+import { Label } from 'ng2-charts';
+import { InformationRequestService } from 'src/app/services/request/information-request.service';
+import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
+import { Store } from '@ngxs/store';
+import { CustomToastrService } from 'src/app/services/helper/custom-toastr.service';
 
 @Component({
-  selector: "app-yearbook-details",
-  templateUrl: "./yearbook-details.component.html",
+  selector: 'app-yearbook-details',
+  templateUrl: './yearbook-details.component.html',
   styleUrls: [`./yearbook-details.component.scss`],
 })
 export class YearbookDetailsComponent extends InformationDetailsComponent {
@@ -23,19 +23,19 @@ export class YearbookDetailsComponent extends InformationDetailsComponent {
   dataLogicLapseOne: ChartDataSets[] = [
     {
       data: [],
-      label: "Diagnóstico - Lógica",
+      label: 'Diagnóstico - Lógica',
     },
   ];
   dataMathLapseOne: ChartDataSets[] = [
     {
       data: [],
-      label: "Diagnóstico - Matemáticas",
+      label: 'Diagnóstico - Matemáticas',
     },
   ];
   dataReadingLapseOne: ChartDataSets[] = [
     {
       data: [],
-      label: "Diagnóstico - Lectura",
+      label: 'Diagnóstico - Lectura',
     },
   ];
   labelLapseOne: Label[] = [];
@@ -43,38 +43,38 @@ export class YearbookDetailsComponent extends InformationDetailsComponent {
   dataLogicLapseTwo: ChartDataSets[] = [
     {
       data: [],
-      label: "Diagnóstico - Lógica",
+      label: 'Diagnóstico - Lógica',
     },
   ];
   dataMathLapseTwo: ChartDataSets[] = [
     {
       data: [],
-      label: "Diagnóstico - Matemáticas",
+      label: 'Diagnóstico - Matemáticas',
     },
   ];
   dataReadingLapseTwo: ChartDataSets[] = [
     {
       data: [],
-      label: "Diagnóstico - Lectura",
+      label: 'Diagnóstico - Lectura',
     },
   ];
 
   dataLogicLapseThree: ChartDataSets[] = [
     {
       data: [],
-      label: "Diagnóstico - Lógica",
+      label: 'Diagnóstico - Lógica',
     },
   ];
   dataMathLapseThree: ChartDataSets[] = [
     {
       data: [],
-      label: "Diagnóstico - Matemáticas",
+      label: 'Diagnóstico - Matemáticas',
     },
   ];
   dataReadingLapseThree: ChartDataSets[] = [
     {
       data: [],
-      label: "Diagnóstico - Lectura",
+      label: 'Diagnóstico - Lectura',
     },
   ];
 
