@@ -1,8 +1,9 @@
-import { Input, Output, EventEmitter, OnChanges } from '@angular/core';
+import { Input, Output, EventEmitter, OnChanges, Directive } from '@angular/core';
 import { MESSAGES } from '../../_components/form-components/shared/constant/validation-messages-list';
 import { ACTION } from '../../../_helpers/text-content/text-crud';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 
+@Directive()
 export abstract class BaseForm implements OnChanges {
 
     /**
