@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ReactiveValidationComponent } from './reactive-validation.component';
 
@@ -6,7 +6,7 @@ describe('ReactiveValidationComponent', () => {
   let component: ReactiveValidationComponent;
   let fixture: ComponentFixture<ReactiveValidationComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ReactiveValidationComponent ]
     })
