@@ -10,13 +10,10 @@ import { FORM_MODALITY } from '../../_shared/abstract-form-mode';
   styles: [],
 })
 export class AdminUserComponent {
-
   constructor(
     private formModeService: FormModeService,
-    private modalService: BsModalService) {
-
-    }
-
+    private modalService: BsModalService
+  ) {}
 
   openRegister() {
     this.formModeService.setMode(FORM_MODALITY.CREATE.value);
@@ -26,5 +23,4 @@ export class AdminUserComponent {
       Object.assign({}, { class: 'modal-xl modal-dialog-centered' })
     );
   }
-
 }
