@@ -1,39 +1,39 @@
-import { Component, OnInit, OnChanges, OnDestroy } from "@angular/core";
+import { Component, OnInit, OnChanges, OnDestroy } from '@angular/core';
 import {
   FormControl,
   Validators,
   FormGroup,
   FormBuilder,
-} from "@angular/forms";
-import { ValidationService } from "src/app/pages/_components/form-components/shared/services/validation.service";
+} from '@angular/forms';
+import { ValidationService } from 'src/app/pages/_components/form-components/shared/services/validation.service';
 
 import {
   NORMAL_TEXT_PATTERN,
   NUMBER_PATTERN,
   EMAIL_PATTERN,
-} from "src/app/pages/_components/form-components/shared/constant/validation-patterns-list";
-import { USER_TYPE } from "src/app/_helpers/convention/user-type";
-import { SchoolUserService } from "src/app/services/user/school-user.service";
-import { CustomToastrService } from "src/app/services/helper/custom-toastr.service";
-import { STATUS } from "src/app/_helpers/convention/status";
-import { HttpEvent, HttpEventType } from "@angular/common/http";
-import { Store, Select } from "@ngxs/store";
+} from 'src/app/pages/_components/form-components/shared/constant/validation-patterns-list';
+import { USER_TYPE } from 'src/app/_helpers/convention/user-type';
+import { SchoolUserService } from 'src/app/services/user/school-user.service';
+import { CustomToastrService } from 'src/app/services/helper/custom-toastr.service';
+import { STATUS } from 'src/app/_helpers/convention/status';
+import { HttpEvent, HttpEventType } from '@angular/common/http';
+import { Store, Select } from '@ngxs/store';
 import {
   SetSchoolUser,
   SchoolUserState,
   UpdateSchoolUser,
-} from "src/app/store/user/school-user.action";
-import { Subscription, Observable, BehaviorSubject, Subject } from "rxjs";
-import { SchoolUser, Coordinate } from "src/app/_models/user/school.model";
-import { Role, DEVNAME_ROLE } from "src/app/_models/permission.model";
-import { RolesState } from "src/app/store/role.action";
-import { BaseForm } from "../../../_shared/base-form";
-import { AddressState } from "src/app/store/_address/address.action";
-import { Statal } from "src/app/_models/address.model";
+} from 'src/app/store/user/school-user.action';
+import { Subscription, Observable, BehaviorSubject, Subject } from 'rxjs';
+import { SchoolUser, Coordinate } from 'src/app/_models/user/school.model';
+import { Role, DEVNAME_ROLE } from 'src/app/_models/permission.model';
+import { RolesState } from 'src/app/store/role.action';
+import { BaseForm } from '../../../_shared/base-form';
+import { AddressState } from 'src/app/store/_address/address.action';
+import { Statal } from 'src/app/_models/address.model';
 
 @Component({
-  selector: "app-schools-users-form",
-  templateUrl: "./schools-users-form.component.html",
+  selector: 'app-schools-users-form',
+  templateUrl: './schools-users-form.component.html',
 })
 export class SchoolsUsersFormComponent
   extends BaseForm
@@ -45,20 +45,20 @@ export class SchoolsUsersFormComponent
   subscription: Subscription;
 
   classTime = [
-    { value: "1", label: "Mañana" },
-    { value: "2", label: "Tarde" },
-    { value: "3", label: "Ambos" },
+    { value: '1', label: 'Mañana' },
+    { value: '2', label: 'Tarde' },
+    { value: '3', label: 'Ambos' },
   ];
 
   typeSchool = [
-    { value: "1", label: "Nacional" },
-    { value: "2", label: "Estadal" },
-    { value: "3", label: "Municipal" },
+    { value: '1', label: 'Nacional' },
+    { value: '2', label: 'Estadal' },
+    { value: '3', label: 'Municipal' },
   ];
 
   backupOldData: SchoolUser;
-  idState = "";
-  idMunicipality = "";
+  idState = '';
+  idMunicipality = '';
 
   showProgress = false;
 
@@ -73,7 +73,7 @@ export class SchoolsUsersFormComponent
     private schoolUserService: SchoolUserService,
     private validationService: ValidationService
   ) {
-    super("una escuela");
+    super('una escuela');
   }
 
   ngOnInit(): void {
@@ -81,75 +81,75 @@ export class SchoolsUsersFormComponent
     this.form.controls.name.setValidators([Validators.required]);
 
     // Data school
-    this.form.addControl("image", new FormControl(null));
-    this.form.addControl("code", new FormControl("", [Validators.required]));
-    this.form.addControl("role", new FormControl());
+    this.form.addControl('image', new FormControl(null));
+    this.form.addControl('code', new FormControl('', [Validators.required]));
+    this.form.addControl('role', new FormControl());
 
     // Data address
     this.form.addControl(
-      "addressCity",
-      new FormControl("", [Validators.required])
+      'addressCity',
+      new FormControl('', [Validators.required])
     );
     this.form.addControl(
-      "addressZoneType",
+      'addressZoneType',
       new FormControl(null, [Validators.required])
     );
     this.form.addControl(
-      "addressZone",
+      'addressZone',
       new FormControl(null, [Validators.required])
     );
 
     // Data principal
-    this.form.addControl("principalFirstName", new FormControl(""));
-    this.form.addControl("principalLastName", new FormControl(""));
-    this.form.addControl("principalEmail", new FormControl(""));
-    this.form.addControl("principalPhone", new FormControl(""));
+    this.form.addControl('principalFirstName', new FormControl(''));
+    this.form.addControl('principalLastName', new FormControl(''));
+    this.form.addControl('principalEmail', new FormControl(''));
+    this.form.addControl('principalPhone', new FormControl(''));
 
     // Data sub principal
     this.form.addControl(
-      "subPrincipalFirstName",
+      'subPrincipalFirstName',
       new FormControl(null, [Validators.pattern(NORMAL_TEXT_PATTERN)])
     );
     this.form.addControl(
-      "subPrincipalLastName",
+      'subPrincipalLastName',
       new FormControl(null, [Validators.pattern(NORMAL_TEXT_PATTERN)])
     );
     this.form.addControl(
-      "subPrincipalEmail",
+      'subPrincipalEmail',
       new FormControl(null, [Validators.pattern(EMAIL_PATTERN)])
     );
     this.form.addControl(
-      "subPrincipalPhone",
+      'subPrincipalPhone',
       new FormControl(null, [Validators.pattern(NUMBER_PATTERN)])
     );
 
     // Data extra school
     this.form.addControl(
-      "nTeachers",
+      'nTeachers',
       new FormControl(null, [Validators.pattern(NUMBER_PATTERN)])
     );
     this.form.addControl(
-      "nAdministrativeStaff",
+      'nAdministrativeStaff',
       new FormControl(null, [Validators.pattern(NUMBER_PATTERN)])
     );
     this.form.addControl(
-      "nLaborStaff",
+      'nLaborStaff',
       new FormControl(null, [Validators.pattern(NUMBER_PATTERN)])
     );
     this.form.addControl(
-      "nStudents",
+      'nStudents',
       new FormControl(null, [Validators.pattern(NUMBER_PATTERN)])
     );
     this.form.addControl(
-      "nGrades",
+      'nGrades',
       new FormControl(null, [Validators.pattern(NUMBER_PATTERN)])
     );
     this.form.addControl(
-      "nSections",
+      'nSections',
       new FormControl(null, [Validators.pattern(NUMBER_PATTERN)])
     );
-    this.form.addControl("schoolShift", new FormControl());
-    this.form.addControl("schoolType", new FormControl());
+    this.form.addControl('schoolShift', new FormControl());
+    this.form.addControl('schoolType', new FormControl());
 
     this.form.updateValueAndValidity();
   }
@@ -160,7 +160,7 @@ export class SchoolsUsersFormComponent
 
     if (this.MODE === this.ACTION.EDIT) {
       this.subscription = this.user$.subscribe((response) => {
-        this.title = "Actualizar usuario escuela";
+        this.title = 'Actualizar usuario escuela';
 
         this.backupOldData = response;
 
@@ -183,29 +183,29 @@ export class SchoolsUsersFormComponent
         this.coordinate.latitude = response.coordinate.latitude;
         this.coordinate.longitude = response.coordinate.longitude;
 
-        this.form.get("password").setValue("");
-        this.form.get("password").clearValidators();
-        this.form.get("password").updateValueAndValidity();
+        this.form.get('password').setValue('');
+        this.form.get('password').clearValidators();
+        this.form.get('password').updateValueAndValidity();
       });
     } else if (this.MODE === this.ACTION.CREATE) {
-      if ("geolocation" in navigator) {
+      if ('geolocation' in navigator) {
         navigator.geolocation.getCurrentPosition((position) => {
           this.coordinate.latitude = position.coords.latitude;
           this.coordinate.longitude = position.coords.longitude;
         });
       }
 
-      this.title = "Registrar usuario escuela";
+      this.title = 'Registrar usuario escuela';
       this.restar();
 
       this.form
-        .get("password")
+        .get('password')
         .setValidators([
           Validators.required,
           Validators.minLength(8),
           Validators.maxLength(8),
         ]);
-      this.form.get("password").updateValueAndValidity();
+      this.form.get('password').updateValueAndValidity();
       this.idState = null;
 
       this.subscription = this.role$.subscribe((response) => {
@@ -236,7 +236,7 @@ export class SchoolsUsersFormComponent
         data.userType = USER_TYPE.SCHOOL.VALUE;
         data.addressZoneType = data.addressZoneType.toString();
 
-        this.toastr.info("Guardando", "Enviando información, espere...");
+        this.toastr.info('Guardando', 'Enviando información, espere...');
 
         this.showProgress = true;
         data = { ...data, coordinate: this.coordinate };
@@ -250,8 +250,8 @@ export class SchoolsUsersFormComponent
               case HttpEventType.Response:
                 this.store.dispatch(new SetSchoolUser(event.body));
                 this.toastr.registerSuccess(
-                  "Registro",
-                  "Usuario escuela registrado"
+                  'Registro',
+                  'Usuario escuela registrado'
                 );
                 this.restar();
                 break;
@@ -264,37 +264,37 @@ export class SchoolsUsersFormComponent
 
             if (err.error.status === 0) {
               this.toastr.error(
-                "Error de datos",
-                "Verifica los datos del formulario"
+                'Error de datos',
+                'Verifica los datos del formulario'
               );
 
               this.toastr.info(
-                "Nota",
-                "Recuerda dar permiso al servicio de Google Maps, Las coordenadas son requeridas."
+                'Nota',
+                'Recuerda dar permiso al servicio de Google Maps, Las coordenadas son requeridas.'
               );
             }
 
             if (err.error.code) {
               this.toastr.error(
-                "Error de datos",
-                "El código del plantel esta en uso"
+                'Error de datos',
+                'El código del plantel esta en uso'
               );
             }
 
             if (err.error.cardId) {
-              if (String(err.error.cardId[0].status) === "5") {
+              if (String(err.error.cardId[0].status) === '5') {
                 this.toastr.error(
-                  "Error de indentidad",
-                  "El documento de identidad ya esta registrado"
+                  'Error de indentidad',
+                  'El documento de identidad ya esta registrado'
                 );
               }
             }
 
             if (err.error.email) {
-              if (String(err.error.email[0].status) === "5") {
+              if (String(err.error.email[0].status) === '5') {
                 this.toastr.error(
-                  "Datos duplicados",
-                  "El correo que se intenta registra ya existe."
+                  'Datos duplicados',
+                  'El correo que se intenta registra ya existe.'
                 );
               }
             }
@@ -303,7 +303,7 @@ export class SchoolsUsersFormComponent
       } else if (this.MODE === this.ACTION.EDIT) {
         let updateData: any = this.form.value;
 
-        if (updateData.password === "" || updateData.password === null) {
+        if (updateData.password === '' || updateData.password === null) {
           delete updateData.password;
         }
 
@@ -321,13 +321,13 @@ export class SchoolsUsersFormComponent
                 new UpdateSchoolUser(this.backupOldData, event)
               );
               this.toastr.updateSuccess(
-                "Actualización",
-                "Usuario actualizado satisfactoriamente"
+                'Actualización',
+                'Usuario actualizado satisfactoriamente'
               );
               this.submitted = false;
-              this.form.get("password").setValue("");
-              this.form.get("password").setValidators([]);
-              this.form.get("password").updateValueAndValidity();
+              this.form.get('password').setValue('');
+              this.form.get('password').setValidators([]);
+              this.form.get('password').updateValueAndValidity();
 
               setTimeout(() => {
                 this.showProgress = false;
@@ -338,37 +338,37 @@ export class SchoolsUsersFormComponent
               console.log(err);
               if (err.error.code) {
                 this.toastr.error(
-                  "Error de datos",
-                  "El código del plantel esta en uso"
+                  'Error de datos',
+                  'El código del plantel esta en uso'
                 );
               }
 
               if (err.error.status === 0) {
                 this.toastr.error(
-                  "Error de datos",
-                  "Verifica los datos del formulario"
+                  'Error de datos',
+                  'Verifica los datos del formulario'
                 );
 
                 this.toastr.info(
-                  "Nota",
-                  "Recuerda dar permiso al servicio de Google Maps, Las coordenadas son requeridas."
+                  'Nota',
+                  'Recuerda dar permiso al servicio de Google Maps, Las coordenadas son requeridas.'
                 );
               }
 
               if (err.error.cardId) {
-                if (String(err.error.cardId[0].status) === "5") {
+                if (String(err.error.cardId[0].status) === '5') {
                   this.toastr.error(
-                    "Error de indentidad",
-                    "El documento de identidad ya esta registrado"
+                    'Error de indentidad',
+                    'El documento de identidad ya esta registrado'
                   );
                 }
               }
 
               if (err.error.email) {
-                if (String(err.error.email[0].status) === "5") {
+                if (String(err.error.email[0].status) === '5') {
                   this.toastr.error(
-                    "Datos duplicados",
-                    "El correo que se intenta registra ya existe."
+                    'Datos duplicados',
+                    'El correo que se intenta registra ya existe.'
                   );
                 }
               }
@@ -395,22 +395,22 @@ export class SchoolsUsersFormComponent
 
   private sanitizeNoRequiredData(data: any) {
     data.subPrincipalFirstName =
-      data.subPrincipalFirstName === "" ? null : data.subPrincipalFirstName;
+      data.subPrincipalFirstName === '' ? null : data.subPrincipalFirstName;
     data.subPrincipalLastName =
-      data.subPrincipalLastName === "" ? null : data.subPrincipalLastName;
+      data.subPrincipalLastName === '' ? null : data.subPrincipalLastName;
     data.subPrincipalPhone =
-      data.subPrincipalPhone === "" ? null : data.subPrincipalPhone;
+      data.subPrincipalPhone === '' ? null : data.subPrincipalPhone;
     data.subPrincipalEmail =
-      data.subPrincipalEmail === "" ? null : data.subPrincipalEmail;
-    data.nTeachers = data.nTeachers === "" ? null : data.nTeachers;
+      data.subPrincipalEmail === '' ? null : data.subPrincipalEmail;
+    data.nTeachers = data.nTeachers === '' ? null : data.nTeachers;
     data.nAdministrativeStaff =
-      data.nAdministrativeStaff === "" ? null : data.nAdministrativeStaff;
-    data.nLaborStaff = data.nLaborStaff === "" ? null : data.nLaborStaff;
-    data.nStudents = data.nStudents === "" ? null : data.nStudents;
-    data.nGrades = data.nGrades === "" ? null : data.nGrades;
-    data.nSections = data.nSections === "" ? null : data.nSections;
-    data.schoolShift = data.schoolShift === "null" ? null : data.schoolShift;
-    data.schoolType = data.schoolType === "null" ? null : data.schoolType;
+      data.nAdministrativeStaff === '' ? null : data.nAdministrativeStaff;
+    data.nLaborStaff = data.nLaborStaff === '' ? null : data.nLaborStaff;
+    data.nStudents = data.nStudents === '' ? null : data.nStudents;
+    data.nGrades = data.nGrades === '' ? null : data.nGrades;
+    data.nSections = data.nSections === '' ? null : data.nSections;
+    data.schoolShift = data.schoolShift === 'null' ? null : data.schoolShift;
+    data.schoolType = data.schoolType === 'null' ? null : data.schoolType;
   }
 
   // -- Event selected rol --
