@@ -8,21 +8,25 @@ import { environment } from 'src/environments/environment';
   providedIn: 'root',
 })
 export class ActivityStrategyService {
-  private readonly ACTIVITY_STRATEGY: string = 'pecasetting/monitoringactivities';
+  private readonly ACTIVITY_STRATEGY: string =
+    'pecasetting/monitoringactivities';
 
   constructor(private httpClient: HttpClient) {}
 
   getActivityStrategy(): Observable<ModelActivityStrategy> {
-    return this.httpClient
-      .get<ModelActivityStrategy>(`${environment.api}${this.ACTIVITY_STRATEGY}`);
+    return this.httpClient.get<ModelActivityStrategy>(
+      `${environment.api}${this.ACTIVITY_STRATEGY}`
+    );
   }
 
   updateActivityStrategy(data: any): Observable<any> {
-    return this.httpClient
-    .post<any>(`${environment.api}${this.ACTIVITY_STRATEGY}`, data, {
-      reportProgress: true,
-      observe: 'events',
-    });
-
+    return this.httpClient.post<any>(
+      `${environment.api}${this.ACTIVITY_STRATEGY}`,
+      data,
+      {
+        reportProgress: true,
+        observe: 'events',
+      }
+    );
   }
 }
