@@ -4,7 +4,7 @@ export const environment = {
   // api: "https://amblema.org/api/",
   // web: "https://amblema.org/",
 
-  api: "http://157.245.131.248:10506/",
-  web: "http://157.245.131.248:10512/",
+  api: 'http://157.245.131.248:10506/',
+  web: 'http://157.245.131.248:10512/',
 
 };
