@@ -7,6 +7,6 @@ export const environment = {
   // api: 'http://186.88.121.11:10505/',
   // -- Web page Amblema --
 
-  api: "http://157.245.131.248:10506/",
-  web: "http://157.245.131.248:10512/",
+  api: 'http://157.245.131.248:10506/',
+  web: 'http://157.245.131.248:10512/',
 };
